# Refactor Object getter/property macros to remove duplications