Extracts def macros to avoid most of the duplication between getter, getter?, property and property? as well as getter! and property! (be they for ivars or class vars).

The def macros are defined in the `Crystal` namespace to avoid polluting the `Object` type with internal, undocumented, macros.

It's still a bit of macro mess, but at least each implementation is no longer duplicated.

The downside is that macros generating calls to other macros is slowing the semantic passes and uses more memory. For example compiling crystal itself:

Before:
```
Semantic (top level):              00:00:00.366039601 ( 190.06MB)
Semantic (new):                    00:00:00.002077768 ( 190.06MB)
Semantic (type declarations):      00:00:00.033126181 ( 198.06MB)
Semantic (abstract def check):     00:00:00.015921465 ( 222.06MB)
Semantic (restrictions augmenter): 00:00:00.010128594 ( 222.06MB)
Semantic (ivars initializers):     00:00:02.880108440 (1152.00MB)
Semantic (cvars initializers):     00:00:00.007371344 (1160.00MB)
```

After:
```
Semantic (top level):              00:00:00.416716450 ( 222.14MB)
Semantic (new):                    00:00:00.002005603 ( 222.14MB)
Semantic (type declarations):      00:00:00.031955767 ( 230.14MB)
Semantic (abstract def check):     00:00:00.013552446 ( 254.14MB)
Semantic (restrictions augmenter): 00:00:00.009634944 ( 254.14MB)
Semantic (ivars initializers):     00:00:03.001631343 (1144.07MB)
Semantic (cvars initializers):     00:00:00.006819680 (1144.07MB)
```