@ysbaddaden 2025-01-28
It's roughly the same for `std_spec`:

```
Semantic (top level):              00:00:00.339730849 ( 190.06MB)
Semantic (new):                    00:00:00.001899277 ( 190.06MB)
Semantic (type declarations):      00:00:00.029891455 ( 198.06MB)
Semantic (abstract def check):     00:00:00.013262059 ( 222.06MB)
Semantic (restrictions augmenter): 00:00:00.008930907 ( 222.06MB)
Semantic (ivars initializers):     00:00:02.681903393 (1152.00MB)
Semantic (cvars initializers):     00:00:00.006673083 (1160.00MB)
```

```
Semantic (top level):              00:00:00.385934620 ( 222.14MB)
Semantic (new):                    00:00:00.001888600 ( 222.14MB)
Semantic (type declarations):      00:00:00.030185148 ( 230.14MB)
Semantic (abstract def check):     00:00:00.013350880 ( 254.14MB)
Semantic (restrictions augmenter): 00:00:00.009861246 ( 254.14MB)
Semantic (ivars initializers):     00:00:02.816699896 (1144.07MB)
Semantic (cvars initializers):     00:00:00.007600426 (1144.07MB)
```