@ysbaddaden 2025-01-28
I simplified how I pass the block to the inner `def_` macros, which requires a bit less memory.

`make .build/crystal`:
- Semantic (new):   0.38s -> 0.42s (+10%) (190MB -> 214MB)
- Semantic (ivars): 2.88s -> 2.96s (+3%) (1152MB -> 1104MB)

`make .build/std_spec`:
- Semantic (new):   5.33s -> 5.43s (+2%) (622MB -> 630MB)
- Semantic (ivars): 0.32s -> 0.05s (-84%) (654MB -> 710MB)

Weirdly, processing the ivars while building the std specs becomes magnitudes faster while using a bit more memory, while building crystal became a little slower but uses less memory.